lib/requirements.txt
nltk
scikit-learn
python-dotenv